+++ src/pages/teams/components/CreateTeamModal.tsx
@@ -1,0 +1,141 @@
+import { useState, useEffect, useRef } from 'react';
+import { type CreateTeamRequest } from '../../../api/teams';
+
+interface CreateTeamModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  onSubmit: (data: CreateTeamRequest) => Promise<void>;
+  isLoading: boolean;
+}
+
+export default function CreateTeamModal({
+  isOpen,
+  onClose,
+  onSubmit,
+  isLoading,
+}: CreateTeamModalProps) {
+  const nameInputRef = useRef<HTMLInputElement>(null);
+  const [formData, setFormData] = useState<CreateTeamRequest>({
+    name: '',
+    description: ''
+  });
+
+  useEffect(() => {
+    if (isOpen && nameInputRef.current) {
+      nameInputRef.current.focus();
+    }
+  }, [isOpen]);
+
+  useEffect(() => {
+    const handleEscape = (e: KeyboardEvent) => {
+      if (e.key === 'Escape' && isOpen && !isLoading) {
+        handleClose();
+      }
+    };
+
+    window.addEventListener('keydown', handleEscape);
+    return () => window.removeEventListener('keydown', handleEscape);
+  }, [isOpen, isLoading]);
+
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+    const { name, value } = e.target;
+    setFormData(prev => ({
+      ...prev,
+      [name]: value
+    }));
+  };
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!formData.name.trim()) return;
+
+    await onSubmit(formData);
+
+    // 성공 시 폼 초기화 (부모 컴포넌트에서 onClose 호출됨)
+    setFormData({ name: '', description: '' });
+  };
+
+  const handleClose = () => {
+    if (!isLoading) {
+      setFormData({ name: '', description: '' });
+      onClose();
+    }
+  };
+
+  if (!isOpen) return null;
+
+  return (
+    <div
+      className="fixed inset-0 bg-black bg-opacity-50 overflow-y-auto h-full w-full z-50 flex items-center justify-center p-4 animate-in fade-in duration-200"
+      onClick={handleClose}
+    >
+      <div
+        className="relative mx-auto w-full max-w-md animate-in zoom-in-95 duration-200"
+        onClick={(e) => e.stopPropagation()}
+      >
+        <div className="bg-white rounded-2xl shadow-2xl">
+          <div className="px-6 py-5 border-b border-gray-200">
+            <h3 className="text-xl font-semibold text-gray-900">
+              새 스터디 팀 만들기
+            </h3>
+          </div>
+          <form onSubmit={handleSubmit} className="p-6 space-y-5">
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                팀 이름 <span className="text-red-500">*</span>
+              </label>
+              <input
+                ref={nameInputRef}
+                type="text"
+                name="name"
+                value={formData.name}
+                onChange={handleInputChange}
+                required
+                className="block w-full px-4 py-2.5 border border-gray-300 rounded-lg shadow-sm focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-colors"
+                placeholder="알고리즘 스터디"
+              />
+            </div>
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-2">
+                설명
+              </label>
+              <textarea
+                name="description"
+                value={formData.description || ''}
+                onChange={handleInputChange}
+                rows={4}
+                className="block w-full px-4 py-2.5 border border-gray-300 rounded-lg shadow-sm focus:ring-2 focus:ring-blue-500 focus:border-blue-500 transition-colors resize-none"
+                placeholder="팀에 대한 간단한 설명을 작성해주세요"
+              />
+            </div>
+            <div className="flex justify-end gap-3 pt-2">
+              <button
+                type="button"
+                onClick={handleClose}
+                disabled={isLoading}
+                className="px-5 py-2.5 text-sm font-medium text-gray-700 bg-white border border-gray-300 rounded-lg hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+              >
+                취소
+              </button>
+              <button
+                type="submit"
+                disabled={isLoading || !formData.name.trim()}
+                className="px-5 py-2.5 text-sm font-medium text-white bg-blue-600 border border-transparent rounded-lg hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+              >
+                {isLoading ? (
+                  <span className="flex items-center gap-2">
+                    <svg className="animate-spin h-4 w-4" viewBox="0 0 24 24">
+                      <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" fill="none" />
+                      <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
+                    </svg>
+                    생성 중...
+                  </span>
+                ) : '생성'}
+              </button>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  );
+}
